requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.21.0=pypi_0
addict=2.4.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
aliyun-python-sdk-core=2.15.1=pypi_0
aliyun-python-sdk-kms=2.16.2=pypi_0
altair=5.3.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
arxiv=2.1.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
babel=2.14.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bert-score=0.3.13=pypi_0
bitsandbytes=0.43.1=pypi_0
bleach=6.1.0=pypi_0
blinker=1.8.0=pypi_0
blis=0.7.11=pypi_0
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
catalogue=2.0.10=pypi_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.16.0=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
confection=0.1.4=pypi_0
contourpy=1.2.1=pypi_0
crcmod=1.7=pypi_0
cryptography=42.0.5=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
datasets=2.18.0=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
deepspeed=0.14.0=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
einops=0.6.1=pypi_0
einops-exts=0.0.4=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fastapi=0.110.3=pypi_0
fastjsonschema=2.19.1=pypi_0
feedparser=6.0.10=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.13.4=pypi_0
fire=0.6.0=pypi_0
fonttools=4.51.0=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
ftfy=6.2.0=pypi_0
func-timeout=4.3.5=pypi_0
gast=0.5.4=pypi_0
gdown=5.1.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gradio=4.16.0=pypi_0
gradio-client=0.8.1=pypi_0
griffe=0.44.0=pypi_0
h11=0.14.0=pypi_0
h5py=3.11.0=pypi_0
hjson=3.1.0=pypi_0
httpcore=0.17.3=pypi_0
httpx=0.24.0=pypi_0
huggingface-hub=0.22.2=pypi_0
idna=3.7=pypi_0
imageio=2.34.1=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
ipykernel=6.29.4=pypi_0
ipython=8.24.0=pypi_0
ipywidgets=8.1.2=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.4.0=pypi_0
json5=0.9.25=pypi_0
jsonpickle=3.0.4=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.1=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.1.8=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.1=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
kaggle=1.6.12=pypi_0
kiwisolver=1.4.5=pypi_0
lagent=0.2.2=pypi_0
langcodes=3.4.0=pypi_0
language-data=1.2.0=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lightning-utilities=0.11.2=pypi_0
llava=1.2.2.post1=pypi_0
llvmlite=0.43.0=pypi_0
lmdeploy=0.4.0=pypi_0
marisa-trie=1.1.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markdown2=2.4.13=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
medmnist=3.0.1=pypi_0
mistune=3.0.2=pypi_0
mmengine=0.10.4=pypi_0
mmengine-lite=0.10.4=pypi_0
modelscope=1.14.0=pypi_0
mpi4py-mpich=3.1.5=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multimedeval=0.1.1=pypi_0
multiprocess=0.70.16=pypi_0
murmurhash=1.0.10=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.3=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.3=pypi_0
ninja=1.11.1.1=pypi_0
nltk=3.8.1=pypi_0
notebook=7.1.3=pypi_0
notebook-shim=0.2.4=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
open-clip-torch=2.24.0=pypi_0
opencv-python=4.9.0.80=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.0.13=h7f8727e_0
orjson=3.10.1=pypi_0
oss2=2.18.4=pypi_0
overrides=7.7.0=pypi_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
patsy=0.5.6=pypi_0
peft=0.9.0=pypi_0
pexpect=4.9.0=pypi_0
phx-class-registry=4.1.0=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py310h06a4308_0
platformdirs=4.2.1=pypi_0
preshed=3.0.9=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=16.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pycparser=2.22=pypi_0
pycryptodome=3.20.0=pypi_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pydeck=0.9.0b1=pypi_0
pydicom=2.4.4=pypi_0
pydub=0.25.1=pypi_0
pygments=2.17.2=pypi_0
pyjwt=2.8.0=pypi_0
pynndescent=0.5.13=pypi_0
pynvml=11.5.0=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=pypi_0
python=3.10.14=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=2.0.7=pypi_0
python-multipart=0.0.9=pypi_0
python-slugify=8.0.4=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.0.2=pypi_0
qtconsole=5.5.1=pypi_0
qtpy=2.4.1=pypi_0
radgraph=0.0.8=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.0=pypi_0
regex=2024.4.16=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.0=pypi_0
ruff=0.4.2=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.23.2=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
send2trash=1.8.3=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.2.2=py310h06a4308_0
sgmllib3k=1.0.0=pypi_0
shellingham=1.5.4=pypi_0
shortuuid=1.0.13=pypi_0
simplejson=3.19.2=pypi_0
six=1.16.0=pypi_0
smart-open=6.4.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=pypi_0
spacy=3.7.4=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.41.2=h5eee18b_0
srsly=2.4.8=pypi_0
stack-data=0.6.3=pypi_0
starlette=0.37.2=pypi_0
statsmodels=0.14.2=pypi_0
streamlit=1.33.0=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
termcolor=2.4.0=pypi_0
terminado=0.18.1=pypi_0
text-unidecode=1.3=pypi_0
thinc=8.2.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.4.24=pypi_0
tiktoken=0.6.0=pypi_0
timm=0.6.13=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.1=pypi_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
torch=2.1.2=pypi_0
torchmetrics=1.3.2=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.37.2=pypi_0
transformers-stream-generator=0.0.5=pypi_0
triton=2.1.0=pypi_0
typer=0.9.4=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024.1=pypi_0
umap-learn=0.5.7=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.29.0=pypi_0
wasabi=1.1.2=pypi_0
watchdog=4.0.0=pypi_0
wavedrom=2.0.3.post3=pypi_0
wcwidth=0.2.13=pypi_0
weasel=0.3.4=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
websockets=11.0.3=pypi_0
wheel=0.41.2=py310h06a4308_0
widgetsnbextension=4.0.10=pypi_0
xtuner=0.1.18=dev_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_0
yapf=0.40.2=pypi_0
yarl=1.9.4=pypi_0
zhipuai=2.1.0.20240521=pypi_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
